+++ docker-compose.yml
@@ -1,0 +1,41 @@
+services:
+  app:
+    build: .
+    container_name: fastapi_app
+    ports:
+      - "9000:9000"
+    depends_on:
+      - db
+    environment:
+      - DB_HOST=postgres_db     
+      - POSTGRES_DB=postgres
+      - POSTGRES_USER=Intern-Project
+      - POSTGRES_PASSWORD=Yash214!
+      - PYTHONPATH=/app
+    restart: always
+    networks:
+      - bridge
+    command: ["poetry", "run", "python3", "intern_project/main.py"]  
+
+  db:
+    image: postgres:latest
+    container_name: postgres_db
+    restart: always
+    environment:
+      DB_HOST: postgres_db
+      POSTGRES_DB: postgres
+      POSTGRES_USER: Intern-Project
+      POSTGRES_PASSWORD: Yash214!
+
+    networks:
+      - bridge
+    ports:
+      - "5432:5432"
+    volumes:
+      - pg_data:/var/lib/postgresql/data
+networks:
+  bridge:
+    driver: bridge
+      
+volumes:
+  pg_data:
