+++ intern_project/database.py
@@ -2,13 +2,15 @@ import psycopg2
 from psycopg2 import extras
 from fastapi import FastAPI
 from contextlib import asynccontextmanager
+import os
 
-# PostgreSQL Database Configuration
-DB_NAME = "postgres"  
-DB_USER = "Intern-Project"
-DB_PASSWORD = "Yash214!"
-DB_HOST = "localhost"
-DB_PORT = "5432"
+# PostgreSQL Database Configuration (uses Docker environment variables)
+DB_NAME = os.getenv("POSTGRES_DB", "postgres") 
+DB_USER = os.getenv("POSTGRES_USER", "Intern-Project")  
+DB_PASSWORD = os.getenv("POSTGRES_PASSWORD", "Yash214!")
+DB_HOST = os.getenv("DB_HOST", "localhost")  
+DB_PORT = os.getenv("DB_PORT", "5432")
+
 
 # Establishing the Connection
 conn = psycopg2.connect(
@@ -46,5 +48,5 @@ async def lifespan(app: FastAPI):
     conn.close()
     print("Shutdown: Database connection closed.")
 
-# Initialize FastAPI App
+# Initialize FastAPI App 
 APP = FastAPI(lifespan=lifespan)
